Add scikit-learn SVM model; better organise & name models
