--- examples/models/dockerfiles/VGG16.Dockerfile
@@ -1,3 +1,0 @@
-FROM rafikiai/rafiki_worker
-
-RUN pip install pillow
