+++ examples/models/image_classification/SkDt.py
@@ -1,0 +1,122 @@
+from sklearn import tree
+import json
+import pickle
+import os
+import base64
+import numpy as np
+
+from rafiki.dataset import load_dataset
+from rafiki.model import BaseModel, InvalidModelParamsException, test_model_class
+
+class SkDt(BaseModel):
+    '''
+    Implements a decision tree classifier on scikit-learn
+    '''
+
+    def get_knob_config(self):
+        return {
+            'knobs': {
+                'max_depth': {
+                    'type': 'int',
+                    'range': [2, 8]
+                },
+                'criterion': {
+                    'type': 'string',
+                    'values': ['gini', 'entropy']
+                },
+            }
+        }
+
+    def init(self, knobs):
+        self._max_depth = knobs.get('max_depth') 
+        self._criterion = knobs.get('criterion') 
+        self._clf = self._build_classifier(
+            self._max_depth,
+            self._criterion
+        )
+        
+    def train(self, dataset_uri):
+        (images, labels) = self._load_dataset(dataset_uri)
+        X = self._prepare_X(images)
+        y = labels
+        self._clf.fit(X, y)
+
+    def evaluate(self, dataset_uri):
+        (images, labels) = self._load_dataset(dataset_uri)
+        X = self._prepare_X(images)
+        y = labels
+        preds = self._clf.predict(X)
+        accuracy = sum(y == preds) / len(y)
+        return accuracy
+
+    def predict(self, queries):
+        X = self._prepare_X(queries)
+        preds = self._clf.predict(X)
+        return preds
+
+    def destroy(self):
+        pass
+
+    def dump_parameters(self):
+        clf_bytes = pickle.dumps(self._clf)
+        clf_base64 = base64.b64encode(clf_bytes).decode('utf-8')
+        return {
+            'clf_base64': clf_base64
+        }
+
+    def load_parameters(self, params):
+        if 'clf_base64' in params:
+            clf_bytes = base64.b64decode(params['clf_base64'].encode('utf-8'))
+            self._clf = pickle.loads(clf_bytes)
+
+    def _prepare_X(self, images):
+        return [np.array(image).flatten() for image in images]
+
+    def _load_dataset(self, dataset_uri):
+        # Here, we use Rafiki's in-built dataset loader
+        return load_dataset(dataset_uri) 
+
+    def _build_classifier(self, max_depth, criterion):
+        clf = tree.DecisionTreeClassifier(
+            max_depth=max_depth,
+            criterion=criterion
+        ) 
+        return clf
+
+
+if __name__ == '__main__':
+    test_model_class(
+        model_class=SkDt,
+        train_dataset_uri='tf-keras://fashion_mnist?train_or_test=train',
+        test_dataset_uri='tf-keras://fashion_mnist?train_or_test=test',
+        queries=[
+            [[0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 3, 1, 0, 0, 7, 0, 37, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 0, 27, 84, 11, 0, 0, 0, 0, 0, 0, 119, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 88, 143, 110, 0, 0, 0, 0, 22, 93, 106, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 53, 129, 120, 147, 175, 157, 166, 135, 154, 168, 140, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 11, 137, 130, 128, 160, 176, 159, 167, 178, 149, 151, 144, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 1, 0, 2, 1, 0, 3, 0, 0, 115, 114, 106, 137, 168, 153, 156, 165, 167, 143, 157, 158, 11, 0], 
+            [0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 3, 0, 0, 89, 139, 90, 94, 153, 149, 131, 151, 169, 172, 143, 159, 169, 48, 0], 
+            [0, 0, 0, 0, 0, 0, 2, 4, 1, 0, 0, 0, 98, 136, 110, 109, 110, 162, 135, 144, 149, 159, 167, 144, 158, 169, 119, 0], 
+            [0, 0, 2, 2, 1, 2, 0, 0, 0, 0, 26, 108, 117, 99, 111, 117, 136, 156, 134, 154, 154, 156, 160, 141, 147, 156, 178, 0], 
+            [3, 0, 0, 0, 0, 0, 0, 21, 53, 92, 117, 111, 103, 115, 129, 134, 143, 154, 165, 170, 154, 151, 154, 143, 138, 150, 165, 43], 
+            [0, 0, 23, 54, 65, 76, 85, 118, 128, 123, 111, 113, 118, 127, 125, 139, 133, 136, 160, 140, 155, 161, 144, 155, 172, 161, 189, 62], 
+            [0, 68, 94, 90, 111, 114, 111, 114, 115, 127, 135, 136, 143, 126, 127, 151, 154, 143, 148, 125, 162, 162, 144, 138, 153, 162, 196, 58], 
+            [70, 169, 129, 104, 98, 100, 94, 97, 98, 102, 108, 106, 119, 120, 129, 149, 156, 167, 190, 190, 196, 198, 198, 187, 197, 189, 184, 36], 
+            [16, 126, 171, 188, 188, 184, 171, 153, 135, 120, 126, 127, 146, 185, 195, 209, 208, 255, 209, 177, 245, 252, 251, 251, 247, 220, 206, 49], 
+            [0, 0, 0, 12, 67, 106, 164, 185, 199, 210, 211, 210, 208, 190, 150, 82, 8, 0, 0, 0, 178, 208, 188, 175, 162, 158, 151, 11], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], 
+            [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]]
+        ]
+    )
