+++ examples/models/image_classification/SkSvm.py
@@ -1,171 +1,106 @@
-import tensorflow as tf
-from tensorflow import keras
-from tensorflow.keras.preprocessing.image import img_to_array, array_to_img
+from sklearn import svm
 import json
+import pickle
 import os
-import tempfile
-import numpy as np
 import base64
-import abc
-from urllib.parse import urlparse, parse_qs 
+import numpy as np
 
 from rafiki.dataset import load_dataset
 from rafiki.model import BaseModel, InvalidModelParamsException, test_model_class
 
-class VGG16TensorflowModel(BaseModel):
+class SkSvm(BaseModel):
+    '''
+    Implements a SVM on scikit-learn
+    '''
 
     def get_knob_config(self):
         return {
             'knobs': {
-                'epochs': {
+                'max_iter': {
                     'type': 'int',
-                    'range': [1, 1]
+                    'range': [10, 10]
                 },
-                'learning_rate': {
+                'kernel': {
+                    'type': 'string',
+                    'values': ['rbf', 'linear']
+                },
+                'gamma': {
+                    'type': 'string',
+                    'values': ['scale', 'auto']
+                },
+                'C': {
                     'type': 'float_exp',
-                    'range': [1e-5, 1e-1]
-                },
-                'batch_size': {
-                    'type': 'int_cat',
-                    'values': [1, 2, 4, 8, 16, 32, 64, 128]
+                    'range': [1e-2, 1e2]
                 }
-            },
-            'root_knobs': ['hidden_layer_units', 'epochs', 'learning_rate', 'batch_size'],
-            'conditional_knobs': {}
+            }
         }
 
     def init(self, knobs):
-        self._batch_size = knobs.get('batch_size')
-        self._epochs = knobs.get('epochs')
-        self._learning_rate = knobs.get('learning_rate')
-
-        self._graph = tf.Graph()
-        self._sess = tf.Session(graph=self._graph)
-
+        self._max_iter = knobs.get('max_iter') 
+        self._kernel = knobs.get('kernel') 
+        self._gamma = knobs.get('gamma') 
+        self._C = knobs.get('C') 
+        self._clf = self._build_classifier(
+            self._max_iter,
+            self._kernel,
+            self._gamma,
+            self._C
+        )
+        
     def train(self, dataset_uri):
         (images, labels) = self._load_dataset(dataset_uri)
-        images = images.reshape(-1, 784)
-        images = np.dstack([images] * 3)
-        images = images.reshape(-1, 28, 28, 3)
-        images = np.asarray([img_to_array(array_to_img(im, scale=False).resize((48,48))) for im in images])
-
-        num_classes = len(np.unique(labels))
-
-        X = [images]
-        y = keras.utils.to_categorical(
-            labels, 
-            num_classes=num_classes
-        )
-
-        with self._graph.as_default():
-            self._model = self._build_model(num_classes)
-            with self._sess.as_default():
-                self._model.fit(
-                    X, 
-                    y, 
-                    epochs=self._epochs, 
-                    batch_size=self._batch_size
-                )
+        X = self._prepare_X(images)
+        y = labels
+        self._clf.fit(X, y)
 
     def evaluate(self, dataset_uri):
         (images, labels) = self._load_dataset(dataset_uri)
-        images = images.reshape(-1, 784)
-        images = np.dstack([images] * 3)
-        images = images.reshape(-1, 28, 28, 3)
-        images = np.asarray([img_to_array(array_to_img(im, scale=False).resize((48,48))) for im in images])
-
-        num_classes = len(np.unique(labels))
-
-        X = [images]
-        y = keras.utils.to_categorical(
-            labels, 
-            num_classes=num_classes
-        )
-
-        preds = self.predict(X)
-
-        accuracy = sum(labels == preds) / len(y)
+        X = self._prepare_X(images)
+        y = labels
+        preds = self._clf.predict(X)
+        accuracy = sum(y == preds) / len(y)
         return accuracy
 
     def predict(self, queries):
-        X = np.array(queries)
-        with self._graph.as_default():
-            with self._sess.as_default():
-                probs = self._model.predict(X)
-                preds = np.argmax(probs, axis=1)
-
+        X = self._prepare_X(queries)
+        preds = self._clf.predict(X)
         return preds
-    
+
     def destroy(self):
-        self._sess.close()
+        pass
 
     def dump_parameters(self):
-        # TODO: Not save to & read from a file 
-
-        # Save whole model to temp h5 file
-        tmp = tempfile.NamedTemporaryFile(delete=False)
-        with self._graph.as_default():
-            with self._sess.as_default():
-                self._model.save(tmp.name)
-        
-        # Read from temp h5 file & encode it to base64 string
-        with open(tmp.name, 'rb') as f:
-            h5_model_bytes = f.read()
-
-        h5_model_base64 = base64.b64encode(h5_model_bytes).decode('utf-8')
-
-        # Remove temp file
-        os.remove(tmp.name)
-
+        clf_bytes = pickle.dumps(self._clf)
+        clf_base64 = base64.b64encode(clf_bytes).decode('utf-8')
         return {
-            'h5_model_base64': h5_model_base64
+            'clf_base64': clf_base64
         }
 
     def load_parameters(self, params):
-        h5_model_base64 = params.get('h5_model_base64', None)
+        if 'clf_base64' in params:
+            clf_bytes = base64.b64decode(params['clf_base64'].encode('utf-8'))
+            self._clf = pickle.loads(clf_bytes)
 
-        if h5_model_base64 is None:
-            raise InvalidModelParamsException()
-
-        # TODO: Not save to & read from a file 
-
-        # Convert back to bytes & write to temp file
-        tmp = tempfile.NamedTemporaryFile(delete=False)
-        h5_model_bytes = base64.b64decode(h5_model_base64.encode('utf-8'))
-        with open(tmp.name, 'wb') as f:
-            f.write(h5_model_bytes)
-
-        # Load model from temp file
-        with self._graph.as_default():
-            with self._sess.as_default():
-                self._model = keras.models.load_model(tmp.name)
-        
-        # Remove temp file
-        os.remove(tmp.name)
+    def _prepare_X(self, images):
+        return [np.array(image).flatten() for image in images]
 
     def _load_dataset(self, dataset_uri):
         # Here, we use Rafiki's in-built dataset loader
         return load_dataset(dataset_uri) 
 
-    def _build_model(self, num_classes):
-        learning_rate = self._learning_rate
-        model = keras.applications.VGG16(
-            include_top=True,
-            input_shape=(48, 48, 3),
-            weights=None, 
-            classes=num_classes
-        )
+    def _build_classifier(self, max_iter, kernel, gamma, C):
+        clf = svm.SVC(
+            max_iter=max_iter,
+            kernel=kernel,
+            gamma=gamma,
+            C=C
+        ) 
+        return clf
 
-        model.compile(
-            optimizer=keras.optimizers.Adam(lr=learning_rate),
-            loss='categorical_crossentropy',
-            metrics=['accuracy']
-        )
-        return model
 
 if __name__ == '__main__':
     test_model_class(
-        model_class=VGG16TensorflowModel,
+        model_class=SkSvm,
         train_dataset_uri='tf-keras://fashion_mnist?train_or_test=train',
         test_dataset_uri='tf-keras://fashion_mnist?train_or_test=test',
         queries=[
